Changed type to long for AssociationTypeIds

+++ HubSpot.NET/Api/Associations/HubSpotAssociationsTypeIds.cs
@@ -5,19 +5,19 @@ public static class HubSpotAssociationsTypeIds
 {
     public static class CompanyAssociationsTypeIds
     {
-        public static int CompanyToPrimaryContact = 2;
-        public static int CompanyToContact = 280;
+        public static long CompanyToPrimaryContact = 2;
+        public static long CompanyToContact = 280;
     }
 }
 
 public static class HubSpotAssociationsTypeIdsExtensions
 {
-    public static bool IsPrimaryContactForCompany(this int associationsTypeId)
+    public static bool IsPrimaryContactForCompany(this long associationsTypeId)
     {
         return associationsTypeId == HubSpotAssociationsTypeIds.CompanyAssociationsTypeIds.CompanyToPrimaryContact;
     }
 
-    public static bool IsContactForCompany(this int associationsTypeId)
+    public static bool IsContactForCompany(this long associationsTypeId)
     {
         return associationsTypeId == HubSpotAssociationsTypeIds.CompanyAssociationsTypeIds.CompanyToContact;
     }
